+++ Source/Pd/PdObject.hpp
@@ -83,7 +83,8 @@ namespace pd
             VuMeter          = 9,
             Comment          = 10,
             AtomNumber       = 11,
-            AtomSymbol       = 12
+            AtomSymbol       = 12,
+            Array            = 13
         };
         
         //! @brief The constructor for an empty Object.
@@ -133,6 +134,8 @@ namespace pd
         void setSymbol(std::string const& value) noexcept;
         
         std::array<int, 4> getBounds() const noexcept override;
+        
+        std::string getArraySymbol() const noexcept;
     private:
 
         //! @brief The constructor for a new Object.